+++ finish/inventory/src/main/java/io/openliberty/guides/inventory/InventoryResource.java
@@ -35,13 +35,18 @@ public class InventoryResource {
     @Path("/{hostname}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getPropertiesForHost(@PathParam("hostname") String hostname) {
+        // Get properties
         Properties props = manager.get(hostname);
+
         if (props == null) {
             return Response.status(Response.Status.NOT_FOUND)
                            .entity("ERROR: Unknown hostname or the system service may " 
                                  + "not be running on " + hostname)
                            .build();
         }
+
+        // Add properties
+        manager.add(hostname, props);
         return Response.ok(props).build();
     }
 
@@ -50,5 +55,4 @@ public class InventoryResource {
     public InventoryList listContents() {
         return manager.list();
     }
-    
 }
