+++ start/inventory/src/main/java/io/openliberty/guides/inventory/InventoryManager.java
@@ -13,29 +13,36 @@
 package io.openliberty.guides.inventory;
 
 import java.util.Properties;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import io.openliberty.guides.inventory.client.SystemClient;
-import io.openliberty.guides.inventory.model.InventoryList;
+import io.openliberty.guides.inventory.model.*;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 @ApplicationScoped
 public class InventoryManager {
 
-    private InventoryList invList = new InventoryList();
+    private List<SystemData> systems = Collections.synchronizedList(new ArrayList<>());
     private SystemClient systemClient = new SystemClient();
 
     public Properties get(String hostname) {
         systemClient.init(hostname, 9080);
-        
-        Properties properties = systemClient.getProperties();
-        if (properties != null) {
-            invList.addToInventoryList(hostname, properties);
-        }
-        return properties;
+        return systemClient.getProperties();
+    }
+
+    public void add(String hostname, Properties systemProps) {
+        Properties props = new Properties();
+        props.setProperty("os.name", systemProps.getProperty("os.name"));
+        props.setProperty("user.name", systemProps.getProperty("user.name"));
+
+        SystemData host = new SystemData(hostname, props);
+        if (!systems.contains(host))
+            systems.add(host);
     }
 
     public InventoryList list() {
-        return invList;
+        return new InventoryList(systems);
     }
-    
 }
