+++ start/inventory/src/main/java/io/openliberty/guides/inventory/model/SystemData.java
@@ -16,19 +16,30 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
-public class InventoryList {
+public class SystemData {
 
-    private List<SystemData> systems;
+    private final String hostname;
+    private final Properties properties;
 
-    public InventoryList(List<SystemData> systems) {
-        this.systems = systems;
+    public SystemData(String hostname, Properties properties) {
+        this.hostname = hostname;
+        this.properties = properties;
     }
 
-    public List<SystemData> getSystems() {
-        return systems;
+    public String getHostname() {
+        return hostname;
     }
 
-    public int getTotal() {
-        return systems.size();
+    public Properties getProperties() {
+        return properties;
+    }
+
+    @Override
+    public boolean equals(Object host) {
+        if (host instanceof SystemData) {
+            return hostname.equals(((SystemData) host).getHostname());
+        }
+        return false;
     }
 }
+
